tab-completion cleanup for pxt.functions
