+++ pixeltable/functions/image.py
@@ -2,8 +2,8 @@ import base64
 
 import PIL.Image
 
-from pixeltable.type_system import ImageType, StringType
 import pixeltable.func as func
+from pixeltable.utils.code import local_public_names
 
 
 @func.udf
@@ -14,3 +14,10 @@ def b64_encode(img: PIL.Image.Image, image_format: str = 'png') -> str:
     img.save(bytes_arr, format=image_format)
     b64_bytes = base64.b64encode(bytes_arr.getvalue())
     return b64_bytes.decode('utf-8')
+
+
+__all__ = local_public_names(__name__)
+
+
+def __dir__():
+    return __all__
