+++ pixeltable/functions/string.py
@@ -1,7 +1,8 @@
 from typing import Any
 
-from pixeltable.type_system import StringType
 import pixeltable.func as func
+from pixeltable.type_system import StringType
+from pixeltable.utils.code import local_public_names
 
 
 @func.udf(return_type=StringType(), param_types=[StringType()])
@@ -11,3 +12,10 @@ def str_format(format_str: str, *args: Any, **kwargs: Any) -> str:
     - {<key>} will be replaced by the corresponding value in kwargs
     """
     return format_str.format(*args, **kwargs)
+
+
+__all__ = local_public_names(__name__)
+
+
+def __dir__():
+    return __all__
